requirements.txt
requests
Pillow
selenium
